src/Analytics/requirements.txt
flask
flask-login
flask-openid
flask-mail
flask-sqlalchemy
sqlalchemy-migrate
flask-whooshalchemy
flask-wtf
flask-babel
guess_language
flipflop
coverage
postgres
Flask-Script
Flask-Migrate
psycopg2
flask-socketio
pytest
pytest-cov
enum34
apscheduler
Flask-Session
git+https://github.com/mobolic/facebook-sdk.git#egg=facebook-sdk
redis
flask-hashing
fuzzywuzzy
fuzzywuzzy[speedup]
numpy
http://download.pytorch.org/whl/torch-0.3.0.post4-cp27-none-macosx_10_6_x86_64.whl
torchvision
